requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
langsmith>=0.2.0
langchain-text-splitters>=0.3.0
langchain-experimental>=0.3.0
tiktoken>=0.5.0
faiss-cpu>=1.7.4
beautifulsoup4>=4.12.0
pydantic>=2.0.0
python-dotenv>=1.0.0
gradio>=4.0.0
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0